evaluation environment implemented.
it expects training based on 19 classes: training index 1-17 for trainable objects. (see changes in train_PixelNet.py)
the indexes -1 and 18 are mapped onto 18.

After prediction all label-18 predictions are mapped back to 255. Afterwards the training ids are mapped back to the label ids. (see predict_Pixelnet.py)

in evaluate.py the evaluation is implemented for the first 5 images of validation set. The predictions are based on a very shortly trained model --> bad results

Moreover, the draw function in predict_Pixelnet.py now uses the same color for each label as in the ground truth files.